refactor version 1 and update Post.tsx link href
